backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
PyJWT>=2.8.0
python-multipart
werkzeug
requests
numpy
reportlab
gtts
deep-translator
moviepy
openai-whisper
